+++ OnlySnarf/util/config.py
@@ -41,10 +41,12 @@ def set_config(args):
     # overwrite with provided args
     for key, value in args.items():
       parsed_config[key] = value
-    # turn strings of booleans into actual booleans
+    # turn strings of booleans into actual booleans, fix lists
     for key, value in parsed_config.items():
       if value == "True" or value == "False":
         parsed_config[key] = bool(value)
+      elif value == '[]':
+        parsed_config[key] = []
   except Exception as e:
     print(e)
   ###############