+++ CHANGELOG.md
@@ -785,12 +785,15 @@
 - finished test_discount process
 - beginning & finished test_expiration
 - updates to clearing post / message text
+- finished test_schedule, test_users, test_poll
+- finished ensuring Message & Post marshmallow schemas properly instantiate w/ Poll & Schedule 
 
 ------------------------------------------------------------------------------------
 
 ## TODO
 
-- finish ensuring Message & Post marshmallow schemas properly instantiate w/ Poll & Schedule 
+- figure out where to put atexit --> Webdriver.exit
+
 
 - fix output location of geckodriver logs